requirements.txt
google-auth==2.16.1
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.79.0
python-dotenv<=1.0.0
